+++ src/lib/types.ts
@@ -1,8 +1,7 @@
 import type { ServerOptions } from 'node:https';
 import type { IncomingMessage } from 'node:http';
 
-import type { JWKWithKid } from './types-internals';
-import type { supportedPkceAlgorithms } from './helpers';
+import type { JWKWithKid, supportedPkceAlgorithms } from './types-internals';
 
 export interface TokenRequest {
   scope?: string;
