+++ test/cli.test.ts
@@ -2,7 +2,9 @@ import { writeFile } from 'node:fs/promises';
 
 import { describe, it, expect, vi } from 'vitest';
 
-import { exec, type Output} from './lib/cli-fake-runner';
+import { shift } from '../src/cli';
+
+import { exec, type Output } from './lib/cli-fake-runner';
 
 vi.mock('fs/promises', () => ({
   writeFile: vi.fn().mockImplementation(() => ''),
@@ -114,6 +116,20 @@ describe('CLI', () => {
 
     expect(res.stdout).toMatch(/^OAuth 2 server listening on http:\/\/.+?:\d+$/m);
     expect(res.stdout).toMatch(/^OAuth 2 issuer is http:\/\/localhost:\d+\/$/m);
+  });
+
+  describe('shift', () => {
+    it('throws on empty array', () => {
+      expect(() => shift([])).toThrow();
+    });
+
+    it('throws on array containing an undefined entry', () => {
+      expect(() => shift([undefined])).toThrow();
+    });
+
+    it('does not throw on valid input', () => {
+      expect(() => shift(['a'])).not.toThrow();
+    });
   });
 });
