+++ test/oauth2-service.test.ts
@@ -4,12 +4,12 @@ import qs from 'node:querystring';
 import { describe, it, expect, beforeAll } from 'vitest';
 import request from 'supertest';
 
-import { OAuth2Issuer, OAuth2Service  } from '../src';
-import type { MutableRedirectUri } from '../src/lib/types';
+import { OAuth2Issuer, OAuth2Service } from '../src';
+import type { MutableRedirectUri } from '../src';
 import {
   createPKCECodeChallenge,
   createPKCEVerifier,
-} from '../src/lib/helpers';
+} from '../src/lib/oauth2-service.pkce';
 
 import * as testKeys from './keys';
 import { verifyTokenWithKey } from './lib/test_helpers';
